requirements.txt
watchdog
pytesseract
Pillow
GitPython
PyGithub
requests
beautifulsoup4
python-dotenv
markdown
